refactor: split render components into per-file modules + fix CI
- Split lib/render/registry.tsx into one file per catalog component
  (components/{section-header,stack,question-button,grid,image}.tsx)
  with a shared CatalogType/stackGap module. registry.tsx is now
  just import + defineRegistry.
- Restrict ImageCreditOverlay links to http/https via URL.protocol
  allowlist. Falls back to a non-clickable label when the sourceUrl
  (model-generated) has an unsafe scheme — prevents javascript:/data:
  XSS via inline image spec blocks (Codex P2).
- Re-run prettier on registry.tsx and tavily.ts (CI Format Check
  was failing on the initial commit).
