+++ components/image-credit-overlay.tsx
@@ -3,6 +3,24 @@
 
 import { displayUrlName } from '@/lib/utils/domain'
 
+/**
+ * Normalize a user- or model-supplied URL to a safe http(s) URL, or return
+ * null if the scheme is not allowed. Prevents XSS via javascript:/data: URLs
+ * embedded in model-generated spec blocks.
+ */
+const sanitizeHttpUrl = (raw: string | undefined): string | null => {
+  if (!raw) return null
+  try {
+    const parsed = new URL(raw)
+    if (parsed.protocol === 'http:' || parsed.protocol === 'https:') {
+      return parsed.toString()
+    }
+    return null
+  } catch {
+    return null
+  }
+}
+
 export const getFaviconUrl = (imageUrl: string): string => {
   try {
     const hostname = new URL(imageUrl).hostname
@@ -27,29 +45,50 @@ export function ImageCreditOverlay({
   title,
   description
 }: ImageCreditOverlayProps) {
-  const linkUrl = sourceUrl || url
+  const safeLink = sanitizeHttpUrl(sourceUrl) ?? sanitizeHttpUrl(url)
   const label = title || description
-  return (
-    <a
-      href={linkUrl}
-      target="_blank"
-      rel="noopener noreferrer"
-      className="absolute bottom-3 left-3 max-w-[80%] bg-black/70 backdrop-blur-sm rounded-xl px-3 py-2.5 flex items-center gap-2.5 no-underline hover:bg-black/80 transition-colors"
-      onClick={e => e.stopPropagation()}
-    >
-      <img
-        src={getFaviconUrl(linkUrl)}
-        alt=""
-        className="size-7 rounded-lg shrink-0"
-      />
+
+  const content = (
+    <>
+      {safeLink && (
+        <img
+          src={getFaviconUrl(safeLink)}
+          alt=""
+          className="size-7 rounded-lg shrink-0"
+        />
+      )}
       <div className="min-w-0 flex-1">
-        <div className="text-white/70 text-xs">{displayUrlName(linkUrl)}</div>
+        {safeLink && (
+          <div className="text-white/70 text-xs">
+            {displayUrlName(safeLink)}
+          </div>
+        )}
         {label && (
           <div className="text-white text-sm font-medium line-clamp-1">
             {label}
           </div>
         )}
       </div>
+    </>
+  )
+
+  const className =
+    'absolute bottom-3 left-3 max-w-[80%] bg-black/70 backdrop-blur-sm rounded-xl px-3 py-2.5 flex items-center gap-2.5 no-underline hover:bg-black/80 transition-colors'
+
+  if (!safeLink) {
+    // Untrusted or missing URL: render a non-clickable label instead of an <a>.
+    return <div className={className}>{content}</div>
+  }
+
+  return (
+    <a
+      href={safeLink}
+      target="_blank"
+      rel="noopener noreferrer"
+      className={className}
+      onClick={e => e.stopPropagation()}
+    >
+      {content}
     </a>
   )
 }
